Remove spaces at the beginning of the server name when doing a Battlelog server search.
I don't know why, but some servers appear to have a bunch of spaces at the beginning of their name.  Is that for the purpose of moving them to the top of search results?  Is that to line them up with something on the server loading screen?  I have no idea.  But, anyway, it appears Battlelog automatically finds no results if you search for the server with all of the spaces at the beginning of its name.  So, now the extra spaces are removed before searching Battlelog for the server.